Renamed toastSavedMessage file to services and renamed link in App.js.
